Refactor codebase to improve modularity and maintainability
The existing frontend codebase consisted of a single, 1500+ line `index.tsx` file that handled all application logic, from state management and API calls to DOM rendering and event handling. This monolithic structure made the code difficult to understand, maintain, and extend.

This refactoring introduces a more modular architecture by:

- **Creating a clear directory structure:** The `src` directory is now organized into `api`, `components`, `core`, `styles`, and `utils` to separate concerns.

- **Extracting core logic:** All core business logic, such as data persistence, time calculations, and utility functions, has been moved into dedicated modules within `src/core` and `src/utils`.

- **Modularizing UI components:** All UI-related logic, including the Quantum Timer, Task List, and all modals, has been encapsulated into its own component module under `src/components`. A `modalManager` was created to handle the centralized opening of modals.

- **Cleaning up the main entry point:** The main `index.tsx` file has been significantly cleaned up. It now primarily acts as an orchestrator, initializing the application state and the various component modules.

This new structure makes the codebase significantly easier to navigate and reason about, laying a solid foundation for future development and bug fixing. The build process has been verified to be successful after these changes.
