+++ src/utils/escapeHtml.ts
@@ -1,0 +1,13 @@
+/**
+ * A simple utility to escape HTML characters and prevent XSS.
+ * @param unsafe The string to escape.
+ * @returns The escaped string.
+ */
+export function escapeHtml(unsafe: string): string {
+    return unsafe
+         .replace(/&/g, "&amp;")
+         .replace(/</g, "&lt;")
+         .replace(/>/g, "&gt;")
+         .replace(/"/g, "&quot;")
+         .replace(/'/g, "&#039;");
+}
